requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.22.0
torch>=2.1.0
torchvision>=0.16.0
pillow
scikit-learn
matplotlib
python-multipart
jinja2
